refact: 카테고리 아이콘을 만드는 로직을 분리합니다.
